feat(dashboard): protect doctor dashboard route and refresh appointments after confirmation
