+++ frontend/src/contexts/AuthContext.tsx
@@ -1,19 +1,16 @@
-import { createContext, useState, useEffect,ReactNode } from "react";
-import { User, AuthContextType } from "./auth.types";
-
-
-
+import { createContext, useState, useEffect, ReactNode } from 'react';
+import { User, AuthContextType } from './auth.types';
 
 const AuthContext = createContext<AuthContextType | undefined>(undefined);
 
 function AuthProvider({ children }: { children: ReactNode }) {
   const [user, setUser] = useState<User | null>(null);
   const [token, setToken] = useState<string | null>(null);
-  const [loading, setLoading]=useState(true);
+  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
-    const storedToken = localStorage.getItem("token");
-    const storedUser = localStorage.getItem("user");
+    const storedToken = localStorage.getItem('token');
+    const storedUser = localStorage.getItem('user');
 
     if (storedToken && storedUser) {
       setToken(storedToken);
@@ -25,19 +22,19 @@ function AuthProvider({ children }: { children: ReactNode }) {
   const login = (userData: User, tokenData: string) => {
     setUser(userData);
     setToken(tokenData);
-    localStorage.setItem("token", tokenData);
-    localStorage.setItem("user", JSON.stringify(userData));
+    localStorage.setItem('token', tokenData);
+    localStorage.setItem('user', JSON.stringify(userData));
   };
 
   const logout = () => {
     setUser(null);
     setToken(null);
-    localStorage.removeItem("token");
-    localStorage.removeItem("user");
+    localStorage.removeItem('token');
+    localStorage.removeItem('user')
   };
 
   return (
-    <AuthContext.Provider value={{ user, token, login, logout , loading}}>
+    <AuthContext.Provider value={{ user, token, login, logout, loading }}>
       {children}
     </AuthContext.Provider>
   );
@@ -45,4 +42,3 @@ function AuthProvider({ children }: { children: ReactNode }) {
 
 export { AuthProvider };
 export default AuthContext;
-
